controller: Move interface_filter to internal package
In order to reduce the exported methods visibility.
