+++ pkg/controllers/internal/filter/interface_filter.go
@@ -1,4 +1,4 @@
-package controllers
+package filter
 
 import (
 	v1 "kubevirt.io/api/core/v1"
